Send setCallCount() queries to the client
setCallCount() is invoked by the optimizer to add derived frequency
information to direct calls which are not tracked by the IProfiler.
This function may create new entries in the global IProfiler at the
JITaaS server which is wrong (the global IProfiler at the server
shouldn't even be used). getCallCount() retrieves the information
from the JITaaS client, so we will never find whatever the optimizer
tried to write. Moreover, setCallCount marks the entry as non-persistent
and the client will refuse to serialize anything marked as non-persistent.

This commit makes the required changes to send the setCallCount() issued
at the server to the client. An optimization is implemented where if the
server finds the desired IProfiler entry in the cache and this entry
has the same 'count' value as the one it wants to write, then no remote
message is sent.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>