Improve String.indexOf performance for first character matches
When String.indexOf was called on Power with a string where the first
character matched the target value, the new vectorized implementation
was actually slower than the old scalar implementation due to the
overhead of using vector instructions. To combat this, the first
character will now be compared using scalar instructions. This causes a
slight performance decrease for later matches, but significantly
improves performance for first character matches.

Signed-off-by: Ben Thomas <ben@benthomas.ca>