Reimplement getMethodFromName around other queries
This way it needs no special treatment in TR_J9SharedCacheVM, since the
required records are naturally created within the two simpler queries.

The callingMethod is now ignored because all callers pass the names of
JCL classes that are loaded by the bootstrap loader. The only
restriction that could be imposed by using the loader associated with
callingMethod would be to potentially "hide" the class, but in that case
the existing code retries with the bootstrap class loader anyway. So now
the bootstrap loader is used on the first attempt.

getMethodFromName is split into two virtual methods, one without this
extra parameter, so that OMR can also remove the parameter without
breaking compatibility.

For TR_ValidateMethodFromClassAndSig, the callingClass ("beholder") is
now optional. Callers of getMethodFromClass do not have to provide a
callingClass, and in particular none is provided in getMethodFromName.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>