Define a guaranteed ID for null in the SVM
This allows ID fields of validation and relocation records to be made
optional. Previously, all ID lookups required as a precondition that the
mapped value is not null, simply due to the fact that null was not
stored in the mappings. Now they will continue to require this by
default, so that the lookup does not unexpectedly return null. In cases
where the value is optional, callers must pass SymOptional to explicitly
allow null.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>