Add flag to toggle assembly syntax while building OpenJ9 JDK8 on OSX
OpenJ9 JDK8 is built on OSX 10.10 Yosemite with Xcode4.6.3. The assembly
syntax is different in this setup.

While building m4 files, -DOSX_JAVA8 is enabled to recognize the setup
for JDK8 on OSX.

gcc version on Xcode4 and Xcode9 is 4.2.1. So, compiler version can't be
used to recognize the setup for JDK8 on OSX.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>