Do not set a call count for bytecodes tracked by IProfiler
The optimizer may try to artificially create IProfiler entries for bytecodes
that it does not track. This should be the case for "invokeStatic" and
"invokeSpecial". However, we see this happening for "invokeVirtual" as well.
In this commit we will add a filter such that setCallCount() will do nothing
for calls that are already tracked by IProfiler.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>