Consolidate TR_RelativeMethodAddress
Remove the duplicated code that writes the TR_RelativeMethodAddress
relocation header information, and consolidate it in one canonical
location.

Additionally, this commit fixes an inconsistncy with the
TR_RelativeMethodAddress relocation record; on the compile run, the
compiler would store additional information into the relocation header.
However, on the load run, the aot infrastructure would not use any of
it; worse still, it would think th size of the relocation header was
different. Thus, this relo record, prior to this commit, was completely
functionally incorrect.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>