Avoid SVM symbol/ID lookups for null
Inlined site validations store a receiverClassID based on thisClass, but
the resulting ID is only used for TR_InlinedInterfaceMethodWithNopGuard
and TR_InlinedAbstractMethodWithNopGuard. For the other relocation
kinds, thisClass can be null, but null is not mapped to any ID. The
result of the lookup in that case is NO_ID (0). At load time, ID 0 is
not found, resulting in a null that signifies absence.

This commit prevents both of these lookups when thisClass is null, in
order to allow both getSymbolFromID() and getIDFromSymbol() to start
requiring as a precondition that the requested key is mapped.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>