Fix bad register allocation inside Power intrinsicIndexOf*
Prior to this change, the code responsible for generating code for
intrinsicIndexOf* calls for Power was not marking zeroRegister as
disallowing r0; however, zeroRegister was being used as the base
register in several lvx instructions. Ultimately, it shouldn't matter
whether zeroRegister is allocated to r0, since the value used would be 0
either way.

However, the local register allocator doesn't know this and would
dutifully try to move zeroRegister out of r0 in the middle of internal
control flow. In the event that another register wasn't available and
the register allocator decided to spill one of the other registers used
in the internal control flow, the generated code would be incorrect.
The register allocator would assume that there was no internal control
flow and would thus not ensure that the register allocations were
consistent for labels.

Signed-off-by: Ben Thomas <ben@benthomas.ca>