Ensure JVMTI return values are filled in
Fix all appropriate JVMTI calls to store return value information in
local variables which are initialized to consistent values, and assign
those values at the end of the function unconditionally to any return
pointer which is not NULL.

This does introduce some extra NULL checks for return pointers, but
should have no noticeable performance impact.

Notably not included in this are parameters which describe a
user-allocated buffer (the buffers may or may not be fully or partially
filled in in error situations).

Fixes: #3398

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>