Deprecate ConcurrentHashMap TM support in the JIT
The Java class support does not exist in OpenJ9 and we've found from
past experimentation that the ConcurrentHashMap transactional memory
support did not provide a measurable benefit.

Given the maintenance cost and the fact that the code has not been
running in the last several years we are dropping the JIT support.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>