Adjust _safeReservePhysicalMemoryValue based on amount of free memory
When allocating scratch memory the JIT will avoid allocating the last
TR::Options::_safeReservePhysicalMemoryValue of free physical memory. This
field is initialized to 50 MB, but it makes sense to initialize it to
smaller values if the JVM is running in containers that are very small.
In this commit TR::Options::_safeReservePhysicalMemoryValue is going
to be initialized to a value proportional to the free memory found at the
start of the JVM, but no more than 32 MB. The "free memory" readout already
takes into account the limits imposed on containers.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>