Reduce number of messages related to inlining
JITaaS part of #3506

- Reduce number of calls to IProfiler_searchForMethodSample.
Override methods of ResolvedJ9Method that access fanin info.
This enables us to get number and weight of callers, total weight and
other bucket weight at the creation of client-side mirror
and cache these values on the server.
The number of callers and their weights might change
during compilation, so cached information will not always be
up to date, but it should be a good enough approximation.
Since fanin information is only used at warm and above opt levels,
do not send this information to the server at levels below warm. [skip
ci]

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>