Share common classes' class chain offset storage
Some system classes are loaded very early and are referred to by most
JIT compilations. With this commit the symbol validation manager has a
list of such classes, termed "well-known classes." It will store a
single copy of their class chain offsets in the shared class cache
(SCC). For each AOT-compiled method using the symbol validation manager,
the emitted relocation data needs just one offset into the SCC in order
to refer to the shared data. This not only deduplicates the offsets; it
also allows many records to be eliminated:

- SystemClassByName records are never needed for well-known classes,

- ClassClass records are now completely unnecessary, and

- ClassByName and ClassFromCP records for well-known classes are made
  unnecessary by checking eagerly that the well-known classes are
  visible to all possible "beholders."

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>