Remove some uses of cannotAttemptOSR and account for OSR point marker
If a node is transmuted or created after ILGen, the doNotProfile flag on
its bytecode info will be set, achieving the same purpose as
cannotAttempOSR without affecting all the nodes with the same bytecode
index.

This commit also accounts for potentialOSRPointHelper calls in
StringPeepHoles. The helper calls in a matched pattern will be deleted
if trees in the pattern are transformed. The helper calls are also taken
into account during the match.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>