Make CH table updates more granular
Create a subclass of `TR_PersistentClassInfo` for the client.
Override every method that modifies class info so that it is marked
as dirty/for removal before calling base method.

This enables us to detect changes to persistent CH table at a finer
granularity. We used to mark class as dirty every time class info
is accessed, whether or not it was modified. This is too conservative
and makes CH table updates larger than necessary.

It also makes code cleaner and less prone to bugs, because we don't need
to think about class hierarchy anymore. Since dirty/removed set gets
updated every time a class info is modified, cases like unloading a subclass of
a loaded class are handled automatically, so we don't need
`markSuperClasesAsDirty` method.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>