Add support for DefaultValue bytecode
Adds an initial implementation for defaultValue. This PR does not
update the verifier, that work will be done in a subsequent PR. 
As a result ValueTypes can only used with -Xverify:none for the 
time being. 

A subsequent PR will be added to provide tests for this bytecode.

This based on Eric's work in 
https://github.com/eclipse/openj9/pull/1235.

Signed-off-by: tajila <atobia@ca.ibm.com>