Replace ClassFromCP with ClassByName for fresh classes
If a class is first encountered by the symbol validation manager as the
result of ClassFromCP, it is necessary to generate both the originally
requested ClassFromCP record and a ClassChain validation record. A
ClassByName with the same beholder has the same results as ClassFromCP
would, and it subsumes ClassChain, so that only ClassByName is needed.
The ClassByName record is larger than the ClassFromCP alone, but smaller
than the two records would have been together.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>