Remove redundant potentialOSRPointHelper calls
An potentialOSRPointHelper call is redundant if it is not reached by a
bad yield point, i.e. gen, in HCRGuardAnalysis

The effect of an OSR point on the dataflow is to kill. If no gen reaches
it, it doesn't matter whether we kill there. Thus removing the redundant
potentialOSRPointHelper calls will not change the result of
HCRGuardAnalysis if we were to repeat it after the removal.

Redundant potentialOSRPointHelper calls will result in more OSR guards
being added, therefore should be removed before OSR guard insertion.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>