Add -XX:DebugLocalMapper to force using the debug stack mapper
The debugLocalMapper (debuglocalmapper.c) is much more conservative
about object lifetimes than the regular local mapper.

The regular mapper walks from the current pc to the end of the method
and only reports locals containing objects as alive if they will be
used as objects again.

The debug local mapper walks from the start of the method to the current
pc and determines if the slot previously contained a live object.  This
is the behaviour typically wanted when debugging to ensure that the
debugger can show you the current object values in locals.

Occasionally, we want to force the use of the debug mapper to test
whether the shortened object lifetimes are affecting poorly written user
code, usually due to the use of finalizers.

This PR adds a new option `-XX:DebugLocalMapper` which uses the debug
local mapper to help diagnose and debug these kinds of issues.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>