Set MACOSX_DEPLOYMENT_TARGET before running ./configure for JDK8 on OSX
If MACOSX_DEPLOYMENT_TARGET is not set, then ./configure fails to check
size of (int *) on OSX when building JDK8. Instead of 64 bits, 0 bits is
returned for the size of (int *). Setting MACOSX_DEPLOYMENT_TARGET
before running ./configure for OpenJ9 JDK8 on OSX resolves this issue.
MACOSX_DEPLOYMENT_TARGET is set to the value of MACOSX_VERSION_MIN
(10.9.0) in JKD8 extensions repo.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>