Fix MethodHandle related bugs
- fix filterArguments, server tried to get and use filter indices
from the known object table, which is not supported in JITaaS.
Use an array of booleans to tell if a filter for an argument has filter.

- FoldHandle_argIndices would put indices in the reverse order on
the client, and then reverse them again on the server, resulting
in the non-reversed order, which is incorrect. Fix by not reversing on
the client.

- `X86PrivateLinkage::buildIndirectDispatch` creates an
`invokeExactJ2IThunk`
every time it can, without regard for whether a thunk is even needed.
This leads to crashes, fix the problem by checking if a thunk is needed
before creating it. Note that baseline doesn't create a new thunk if it
already exists, but JITaaS will recreate it anyway, because client might
not have it. [skip ci]

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>