Validate class chain for defining class of static
Since the defining class of a static doesn't escape from the point at
which we request the validation, we're able to get away with checking
only the ROM class, rather than the entire class chain. However, there
are a few issues with this:

- It's different from every other class validation.

- We may need to validate the class chain anyway, if the same class is
  encountered in any other way during compilation, in which case the ROM
  class validation is redundant.

Furthermore, it's unlikely that validating the class chain will cause
many additional load failures.

As such, queries for the defining class of a static will now request a
class chain validation for that class.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>