Eliminate Recompilation_getExistingMethodInfo messages
Persistent method info is initialized in Recompilation
object, which is created for every compilation
that can be recompiled. Thus, it makes sense to send this
method info as a part of compile request.
Once server receives it, method info is allocated and stashed in
`TR::CompilationInfoPerThreadRemote`, until it is time to create a recompilation object.
[skip ci]

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>