Reload evacuated object in the hardware read barrier helper
In the case of the unlikely event that the call to J9ReadBarrier fails
to evacuate the object we may end up in an infinite loop because the
hardware read barrier helper currently returns back into the JIT
mainline to re-execute the guarded load instruction.

The JIT was under the assumption that J9ReadBarrier will always succeed
which may not be the case if we completely run out of space on the heap
to evacuate the object. In such a scenario the GC must reach a
synchronization point and issue a full global collection, but to do
that we must be able to return to the JIT mainline and load the
non-evacuated object. This means we cannot re-execute the guarded load
instruction in the JIT mainline.

To avoid this issue we reload the evacuated object in the hardware read
barrier helper and jump past the guarded load instruction upon exiting
the helper.

As a side-effect of this change we can no longer do constrained TX with
hardware concurrent scavenge because the J9ReadBarrier can fail and we
cannot get out of the CTX loop if that happens because the execution
will resume at the beginning of the transaction.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>

Disable CTX paths in JIT for concurrent scavenge

We cannot handle CTX for concurrent scavenge with hardware support
because the call to evacuate an object via J9ReadBarrier may not
succeed if the heap happens to be full at that time.

With CTX the transaction abort will resume at the transaction begin but
the read barrier handler needs to skip past the guarded load which
initially triggered the guarded storage event, otherwise we will simply
get into an infinite loop of trying to evacuate an object for which
there is no space to evacuate to.

Because this is not possible we must disable CTX in this scenario.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>

Standardize hardware/software read barrier helper function names

Introduce consistency in the naming convention for the two read barrier
handlers; software and hardware.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>