FFI: fix bug on z/os
FFISYS was calling ffi_prep_args to perform some special handling of the
argument types. As part of this ffi_prep_args was returning a list of
argument types  which was store in an array inside the ffi_prep_args stack
frame

However on z/os, it is possible that child stack frames can be reclaimed
after the child returns. Solution is to make FFISYS process the arg types
itself, and the data is no longer returned from ffi_prep_args

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>