Use correct byte code index on asynccheck
Ilgen may generate an if compare node to combine a compare bytecode and
its suceeding conditional branch bytecode. If an asynccheck node is
generated before the if compare node, it should have the bytecode index
of the compare bytecode. This is important in OSR because asynccheck is
an OSR point whose bytecode index represents a transition target. Having
it wrong will result in the interpreter resuming execution from a wrong
place after an OSR transition.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>