Preventing out-of-order updates in JITaaS
The client sends class unloading and CHTable updates with each compilation
request and the JITaaS server needs to apply those updates in the order
they were generated. Unfortunatelly, the updates can arrive out-of-order
at the server and this can lead to functional problems. The solution
implemented by this commit uses a sequencing scheme based on ticket numbers.

The client obtains a sequence number (a ticket) that defines the update
and attaches it to the compilation request.
The server obtains a pointer to the clientSession and then enters a
critical section protected by the "sequencingMonitor" where compilation
requests are reordered. The clientSession maintains an "expectedSeqNo"
variable and only a request with a matching sequence number can pass
through. Any request with a larger sequence number is put onto a
waiting list (we say that handling threads are parked). Waiting is done
on a monitor associated with the entry (this scheme allows us to pick
which threads to wake up).
Any request with a smaller sequence number is cancelled (this can happen
at the very beginning when a later compilation request arrives first and
creates the clientSession).
The thread that goes through applies its updates to the CHTable and the
hashtable with ramClass-->romClass mappings and then increments
expectedSeqNo to allow the next thread to pass. Also, it needs to notify
any potential thread that is parked waiting for its turn.
The first thread that goes through the sequencing critical section and
finds an empty CHTable is responsible for fetching the entire CHTable
from the client (this is done only once).

A big complication is introduced by messages that are lost. If a message
never arrives, the threads that are parked are going to time-out, but
since the lost message could carry crucial information, these threads
cannot continue safely. In order to avoid cancelling these compilation
requests, when the thread at the front of the waiting list times-out,
it waits for any active threads (threads that actually perform
compilations) to drain out, it clears the caches for this clientSession
and then it changes the expectedSeqNo to allow itself to go through the
sequencing critical section. Parked threads are notified as before.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>