Relocate the inlined site table when using the SVM
When using the SVM, by the time the AOT infrastructure gets to applying
relocations, the validation records have guaranteed equivalency in terms
of all relevant assumptions made about the environment. Therefore, an
inlined site relocation can fail because it was forced to via an option,
or the class the guard checks has been extended.

Currently, this will patch the guard and leave the inlined site entry in
the inlined site table in the metadata untouched. However, this can
result in functionally incorrect code. This is because, other
relocations require going to the inlined site table to materialize the
value for the relocation; for example, in order to relocate a pointer to
a static variable. Failing to update the inlined site entry can either
cause the AOT infrastructure to crash, or can result in the relocation
getting ignored.

This commit ensures that, when using the SVM, if an inlined site
relocation fails, the inline site entry in the inlined site table is
relocated.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>