Replace javaVM attribute access with a front-end call
Compiler code should avoid using attributes of javaVM,
and use front-end instead, because it makes code compatible with JITaaS.
In this case, accessing javaVM->booleanArrayClass will lead to bugs on
JITaaS server, because classes from the server VM will be accessed, but
for compilation we need classes from the client.
On the other hand, using front-end call is fine, because JITaaS subclasses `TR_J9VM`,
and will fetch classes from the client.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>