Update cfdump to handle new value type bytecodes
Adds cfdump support for the new bytecodes
- 203, defaultvalue
- 204, withvalue

cfdump assumes unkown bytecodes have a sizeof 1 byte. As these bytecodes
are 3 bytes in size, cases need to be added for them in cfdump.

In addition, more information was added to cfdump in case of future
unimplemented bytecodes. Previously, unimplemented bytecodes were
printed as `JBunimplmented`. Now, the bytecode number is appended, as in
`JBunimplmented_204`.

Signed-off-by: Andrew Crowther <acrowthe3388@gmail.com>