Swap Inlining and methodhandle inlining
The old strategy does methodhandle inlining before normal inlining and
only JSR292 callsites in the outer most method are inlined. A lot of
methodhandle inlining opportunities are missed because of this. This
change swaps inlining and methodhandle inlining to do normal inlining
first to expose the JSR292 in the inlined callees as well.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>