Fix rare VM shutdown assertion
If VM startup fails to complete, but gets far enough to have reported VM
start, then VM shutdown must be reported. The hook caller for this case
was calling the hook in the JNI context, whereas the other two callers
have entered the VM and released VM access. Make the startup failure
case consistent with the other callers.

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>