Eliminate IProfiler::_maxCount
IProfiler::_maxCount is a value set by the optimizer when it calls
"setBlockAndEdgeFrequencies" and is reset at the beginning of a
compilation. The problem is that this value is shared by all compilation
threads so one thread could be setting it to some value while another
thread could be resetting it. The good news is that _maxCount appears
to be unused so it can be deleted.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>