Cache IProfiler info for interpreted methods
Previously, IProfiler would only cache profiling information
for methods that have already been compiled, and for the method currently being compiled.
This change starts caching info for interpreted methods in a per-compilation cache.
When a server requests info for a particular bytecode, client sends profiling info
for the entire method, which is then cached until the end of the compilation.
Even though IProfiler data might get updated during compilation,
the hope is that the change will be small enough to not affect performance.
[skip ci]

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>