Add tests for bytecode verification
Presently, the ValueTypeTests do not verify the bytecodes used. This is
due to the invalid nature of some of the tests, which is used to test
runtime failures or to get around presently unimplemented features.

A seperate test file has been created which will undergo bytecode
verification. This required modifying the bytecode generator to avoid
invalid bytecodes being sent to the verifier.

Signed-off-by: Andrew Crowther <acrowthe3388@gmail.com>