Add infrastructure to retry AOT loads
Currently, when an AOT load fails, the compiler will perform a JIT
compilation. An AOT compilation is generally performed with count=1000,
bcount=250, but an AOT load is generally performed with scount=20. This
means that there will likely be AOT load failures caused by validation
failures which occur because what's being validated isn't resolved yet.
This commit provides the infrastructure to postpone the AOT load (that
fail under specific conditions likely caused by the low scount) and
retry the load some time later.

The general design is as follows. The first time an AOT load fails,
the J9Method is added to a map. The key is the J9Method pointer itself,
and the value is the new invocation count. The extra field of the
J9Method is then updated with the new invocation count, and the
compilation ends. On each subsequent load failure, the invocation count
doubles, up to half of the initial default count/bcount for that
method; the value in the map is used for this purpose and updated
accordingly. The purpose of this is to ensure that the total invocation
count set for this method does not exceed its default count/bcount. If
the count cannot be increased anymore, the method is JIT compiled.

In order to minimize the cost of querying the map (since the map is
only required for AOT load failures), a method is only removed
from the map when the compiler is unable to increase the invocation
count. This is ok because the map is only checked during an AOT load
failure.

Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>