Provide ordering to CHTable updates
The client will send CHTable updates to the JItaaS server, but these
updates need to be applied in the order they were generated at the
client. This commit moves these CHTable updates in the sequencing
region ensuring they are applied in the correct order. More
specifically, each CHTable update encoded at the client will be tagged
with a sequence number which is communicated to the server. The server
logic is reponsible for applying these updates in the order given by
the client sequence number.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>